fix bug

+++ app/controller/project.js
@@ -105,7 +105,7 @@ const Project = class extends Controller {
 		if (!data) return this.throw(500, "记录创建失败");
 		const project = data.get({plain:true});
 
-		if (params.type == PROJECT_TYPE_PARACRAFT) {
+		if (params.type == PROJECT_TYPE_PARACRAFT && params.isCreateWorld) {
 			const ok = await this.createWorld(project);
 			if (!ok) {
 				await this.model.projects.destroy({where:{id:project.id}});
